+++ database/migrations/2016_01_01_05_create_settings_table.php
@@ -13,7 +13,7 @@ class CreateSettingsTable extends Migration
     public function up()
     {
         Schema::create('settings', function (Blueprint $table) {
-            $table->integer('admin_id')->unsigned()->primary();
+            $table->bigInteger('admin_id')->unsigned()->primary();
             $table->integer('year');
 
             $table->foreign('admin_id')->references('student_id')->on('users');
